process_icell8.py: also move functions to icell8.pipeline (where they are needed).
